backend/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.11.7
supabase==2.17.0
python-dotenv==1.1.1
uvicorn==0.35.0
anthropic==0.40.0
httpx==0.27.0
google-generativeai==0.8.3
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-env==1.1.5
